backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "richtato"
version = "0.1.0"
description = "Richtato project"
authors = [
    { name = "brotatokuo", email = "kuoalan9@example.com" }
]

dependencies = [
    "Django==5.1",
    "djangorestframework==3.14.0",
    "django-cors-headers==4.3.1",
    "psycopg2-binary==2.9.7",
    "python-dotenv==1.0.0",
    "colorama==0.4.6",
    "loguru==0.7.2",
    "gunicorn==21.2.0",
    "pandas==2.2.0",
    "numpy==1.26.4",
    "openai==1.3.0",
    "plotly==5.17.0",
    "drf-yasg==1.21.7",
    "fuzzywuzzy==0.18.0",
    "openpyxl==3.1.5",
    "python-Levenshtein==0.26.1",
    "requests==2.32.3",
    "typing-extensions==4.12.2",
    "pdfplumber==0.11.4",
    "pytest==7.4.0",
    "pytest-django==4.5.2",
]


[tool.setuptools.packages.find]
where = ["."]  # Look in the root directory
include = ["richtato*"]  # Include richtato and all its subpackages

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "richtato.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["apps"]
addopts = "-v --tb=short"
pythonpath = ["."]
